Update d user data in Mapper.kt for id type change.

+++ src/main/kotlin/onlytrade/app/login/data/user/mapper/Mapper.kt
@@ -4,7 +4,7 @@ import onlytrade.app.login.data.user.dao.UserDAO
 import onlytrade.app.viewmodel.login.repository.data.db.User
 
 fun UserDAO.toModel(): User = User(
-    id = id.value,
+    id = id.value.toLong(),
     name = usersName,
     password = password,
     phone = phone,
